+++ src/Sequence/Primitive.php
@@ -47,12 +47,6 @@ final class Primitive implements Implementation
         return \count($this->values);
     }
 
-    #[\Override]
-    public function count(): int
-    {
-        return $this->size();
-    }
-
     /**
      * @return Iterator<T>
      */