+++ src/Sequence/Snap.php
@@ -51,12 +51,6 @@ final class Snap implements Implementation
         return $this->memoize()->size();
     }
 
-    #[\Override]
-    public function count(): int
-    {
-        return $this->size();
-    }
-
     /**
      * @return Iterator<T>
      */