+++ tests/Sequence/PrimitiveTest.php
@@ -24,7 +24,6 @@ class PrimitiveTest extends TestCase
     public function testSize()
     {
         $this->assertSame(2, (new Primitive([1, 1]))->size());
-        $this->assertSame(2, (new Primitive([1, 1]))->count());
     }
 
     public function testIterator()
@@ -136,7 +135,7 @@ class PrimitiveTest extends TestCase
 
         $this->assertSame([1, 2, 3, 4], \iterator_to_array($sequence->iterator()));
         $this->assertInstanceOf(Map::class, $groups);
-        $this->assertCount(2, $groups);
+        $this->assertSame(2, $groups->size());
         $this->assertSame([2, 4], $this->get($groups, 0)->toList());
         $this->assertSame([1, 3], $this->get($groups, 1)->toList());
     }